# Feat/optimize encode