### Motivation:

Avoid some memory copy

### Modification:

1. Avoid pd.toBytesArray(), use stream writeTo
2. Avoid memory copy with only 3 bytes on 'RESERVED' of header

### Result:

the perf test on my mac：

```
V1 codec cost:1090 ms.
Total log size:1205400000 bytes.
V2 codec cost:919 ms.
Total log size:1183706496 bytes.

V1 codec cost:1015 ms.
Total log size:1205400000 bytes.
V2 codec cost:994 ms.
Total log size:1183706496 bytes.

V1 codec cost:1237 ms.
Total log size:1205400000 bytes.
V2 codec cost:911 ms.
Total log size:1183706496 bytes.
```